# controllers/security_controller.go
// Package controllers 安全管理控制器
package controllers

import (
	"time"

	"go_wails_project_manager/config"
	"go_wails_project_manager/response"

	"github.com/gin-gonic/gin"
)

// SecurityStatus 安全状态响应
type SecurityStatus struct {
	RateLimiterActive bool        `json:"rate_limiter_active" example:"true"`
	Config            interface{} `json:"config"`
	Uptime            int64       `json:"uptime" example:"3600"`
	ThreatLevel       string      `json:"threat_level" example:"low"`
}

// BlockedIP 被封禁IP信息
type BlockedIP struct {
	IP        string `json:"ip" example:"192.168.1.100"`
	Reason    string `json:"reason" example:"恶意攻击"`
	BlockedAt int64  `json:"blocked_at" example:"1640995200"`
	ExpiresAt int64  `json:"expires_at" example:"1640998800"`
	Duration  int64  `json:"duration" example:"3600"`
}

// IPStats IP统计信息
type IPStats struct {
	IP            string `json:"ip" example:"192.168.1.1"`
	RequestCount  int64  `json:"request_count" example:"150"`
	LastAccess    int64  `json:"last_access" example:"1640995200"`
	IsWhitelisted bool   `json:"is_whitelisted" example:"false"`
	IsBlocked     bool   `json:"is_blocked" example:"false"`
}

// ConnectionStats 连接统计
type ConnectionStats struct {
	TotalConnections  int64            `json:"total_connections" example:"25"`
	ActiveConnections int64            `json:"active_connections" example:"10"`
	IPConnections     map[string]int64 `json:"ip_connections"`
	TopIPs            []IPStats        `json:"top_ips"`
}

// SecurityController 安全控制器
type SecurityController struct{}

// NewSecurityController 创建安全控制器
func NewSecurityController() *SecurityController {
	return &SecurityController{}
}

// GetStatus 获取安全状态
// @Summary 获取系统安全状态
// @Description 获取当前安全中间件运行状态、配置信息和威胁等级
// @Tags 安全管理
// @Accept json
// @Produce json
// @Success 200 {object} response.Response{data=SecurityStatus} "获取成功"
// @Failure 500 {object} response.Response "获取失败"
// @Router /api/security/status [get]
func (sc *SecurityController) GetStatus(c *gin.Context) {
	response.Success(c, gin.H{
		"rate_limiter_active": true,
		"config":              config.GetSecurityConfig(),
	})
}

// GetBlockedIPs 获取被封禁IP列表
// @Summary 获取被封禁IP列表
// @Description 获取当前被封禁的IP地址列表，包括封禁原因和到期时间
// @Tags 安全管理
// @Accept json
// @Produce json
// @Param page query int false "页码，默认1" default(1)
// @Param limit query int false "每页数量，默认20" default(20)
// @Success 200 {object} response.Response{data=[]BlockedIP} "获取成功"
// @Failure 500 {object} response.Response "获取失败"
// @Router /api/security/blocked-ips [get]
func (sc *SecurityController) GetBlockedIPs(c *gin.Context) {
	var blockedList []map[string]interface{}
	// 获取当前黑名单信息
	response.Success(c, gin.H{
		"blocked_ips": blockedList,
		"count":       len(blockedList),
	})
}

// UnblockIP 解封IP地址
// @Summary 手动解封IP地址
// @Description 将指定IP地址从黑名单中移除，恢复其正常访问权限
// @Example ip: 192.168.1.100
// @Tags 安全管理
// @Accept json
// @Produce json
// @Param ip path string true "IP地址（示例: 192.168.1.100）"
// @Success 200 {object} response.Response "解封成功"
// @Failure 400 {object} response.Response "IP地址格式错误"
// @Failure 404 {object} response.Response "IP未被封禁"
// @Failure 500 {object} response.Response "解封失败"
// @Router /api/security/unblock/{ip} [post]
func (sc *SecurityController) UnblockIP(c *gin.Context) {
	ip := c.Param("ip")

	// TODO: 实现 RemoveFromBlacklist 函数
	// RemoveFromBlacklist(ip)
	response.SuccessWithMsg(c, "IP解封成功", gin.H{
		"ip": ip,
	})
}

// GetIPStats 获取IP统计信息
// @Summary 获取IP访问统计
// @Description 获取各IP地址的访问次数、最后访问时间等统计信息
// @Tags 安全管理
// @Accept json
// @Produce json
// @Param page query int false "页码，默认1" default(1)
// @Param limit query int false "每页数量，默认50" default(50)
// @Param sort query string false "排序方式：request_count(请求数) | last_access(最后访问)" default("request_count")
// @Success 200 {object} response.Response{data=[]IPStats} "获取成功"
// @Failure 500 {object} response.Response "获取失败"
// @Router /api/security/ip-stats [get]
func (sc *SecurityController) GetIPStats(c *gin.Context) {
	// 返回统计信息
	response.Success(c, gin.H{
		"message": "安全统计数据",
		"count":   0,
	})
}

// BlockIP 封禁IP地址
// @Summary 手动封禁IP地址
// @Description 将指定IP地址添加到黑名单，阻止其访问系统
// @Tags 安全管理
// @Accept json
// @Produce json
// @Param ip path string true "IP地址（示例: 192.168.1.100）"
// @Param duration query int false "封禁时长(秒)，默认3600" default(3600)
// @Param reason query string false "封禁原因" default("手动封禁")
// @Success 200 {object} response.Response{data=BlockedIP} "封禁成功"
// @Failure 400 {object} response.Response "IP地址格式错误或已被封禁"
// @Failure 500 {object} response.Response "封禁失败"
// @Router /api/security/block/{ip} [post]
func (sc *SecurityController) BlockIP(c *gin.Context) {
	ip := c.Param("ip")
	durationSec := c.DefaultQuery("duration", "3600")
	reason := c.DefaultQuery("reason", "手动封禁")

	var duration time.Duration
	if d, err := time.ParseDuration(durationSec + "s"); err == nil {
		duration = d
	} else {
		duration = time.Hour
	}

	// TODO: 实现 AddToBlacklist 函数
	// AddToBlacklist(ip, duration)
	response.SuccessWithMsg(c, "IP封禁成功", gin.H{
		"ip":       ip,
		"duration": duration.String(),
		"reason":   reason,
	})
}

// AddToWhitelist 添加IP到白名单
// @Summary 添加IP到白名单
// @Description 将IP地址添加到白名单，该IP将绕过所有安全检查
// @Tags 安全管理
// @Accept json
// @Produce json
// @Param ip path string true "IP地址（示例: 192.168.1.1）"
// @Success 200 {object} response.Response "添加成功"
// @Failure 400 {object} response.Response "IP地址格式错误或已在白名单"
// @Failure 500 {object} response.Response "添加失败"
// @Router /api/security/whitelist/{ip} [post]
func (sc *SecurityController) AddToWhitelist(c *gin.Context) {
	ip := c.Param("ip")
	// TODO: 实现 AddToWhitelist 函数
	// AddToWhitelist(ip)
	response.SuccessWithMsg(c, "IP已添加到白名单", gin.H{
		"ip": ip,
	})
}

// RemoveFromWhitelist 从白名单移除IP
// @Summary 从白名单移除IP
// @Description 将IP地址从白名单中移除，该IP将重新受到安全检查
// @Tags 安全管理
// @Accept json
// @Produce json
// @Param ip path string true "IP地址（示例: 192.168.1.1）"
// @Success 200 {object} response.Response "移除成功"
// @Failure 400 {object} response.Response "IP地址格式错误"
// @Failure 404 {object} response.Response "IP不在白名单中"
// @Failure 500 {object} response.Response "移除失败"
// @Router /api/security/whitelist/{ip} [delete]
func (sc *SecurityController) RemoveFromWhitelist(c *gin.Context) {
	ip := c.Param("ip")
	// TODO: 实现 RemoveFromWhitelist 函数
	// RemoveFromWhitelist(ip)
	response.SuccessWithMsg(c, "IP已从白名单移除", gin.H{
		"ip": ip,
	})
}

// GetConnections 获取连接统计
// @Summary 获取实时连接统计
// @Description 获取当前活跃连接统计信息，用于DDoS监控和流量分析
// @Tags 安全管理
// @Accept json
// @Produce json
// @Success 200 {object} response.Response{data=ConnectionStats} "获取成功"
// @Failure 500 {object} response.Response "获取失败"
// @Router /api/security/connections [get]
func (sc *SecurityController) GetConnections(c *gin.Context) {
	// 获取连接统计信息
	response.Success(c, gin.H{
		"total_connections": 0,
		"ip_connections":    map[string]int{},
	})
}