# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "媒体管理器 API 文档，包括媒体库管理、媒体文件管理等功能",
        "title": "媒体管理器 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:23347",
    "basePath": "/",
    "paths": {
...
        "/api/security/block/{ip}": {
            "post": {
                "description": "将指定IP地址添加到黑名单，阻止其访问系统",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "安全管理"
                ],
                "summary": "手动封禁IP地址",
                "parameters": [
                    {
                        "type": "string",
                        "description": "IP地址（示例: 192.168.1.100）",
                        "name": "ip",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 3600,
                        "description": "封禁时长(秒)，默认3600",
                        "name": "duration",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "\"手动封禁\"",
                        "description": "封禁原因",
                        "name": "reason",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "封禁成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/controllers.BlockedIP"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "IP地址格式错误或已被封禁",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "封禁失败",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/security/blocked-ips": {
            "get": {
                "description": "获取当前被封禁的IP地址列表，包括封禁原因和到期时间",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "安全管理"
                ],
                "summary": "获取被封禁IP列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码，默认1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页数量，默认20",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/controllers.BlockedIP"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "获取失败",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/security/connections": {
            "get": {
                "description": "获取当前活跃连接统计信息，用于DDoS监控和流量分析",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "安全管理"
                ],
                "summary": "获取实时连接统计",
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/controllers.ConnectionStats"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "获取失败",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/security/ip-stats": {
            "get": {
                "description": "获取各IP地址的访问次数、最后访问时间等统计信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "安全管理"
                ],
                "summary": "获取IP访问统计",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码，默认1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 50,
                        "description": "每页数量，默认50",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "\"request_count\"",
                        "description": "排序方式：request_count(请求数) | last_access(最后访问)",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/controllers.IPStats"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "获取失败",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/security/status": {
            "get": {
                "description": "获取当前安全中间件运行状态、配置信息和威胁等级",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "安全管理"
                ],
                "summary": "获取系统安全状态",
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/controllers.SecurityStatus"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "获取失败",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/security/unblock/{ip}": {
            "post": {
                "description": "将指定IP地址从黑名单中移除，恢复其正常访问权限",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "安全管理"
                ],
                "summary": "手动解封IP地址",
                "parameters": [
                    {
                        "type": "string",
                        "description": "IP地址（示例: 192.168.1.100）",
                        "name": "ip",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "解封成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "IP地址格式错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "IP未被封禁",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "解封失败",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/security/whitelist/{ip}": {
            "post": {
                "description": "将IP地址添加到白名单，该IP将绕过所有安全检查",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "安全管理"
                ],
                "summary": "添加IP到白名单",
                "parameters": [
                    {
                        "type": "string",
                        "description": "IP地址（示例: 192.168.1.1）",
                        "name": "ip",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "添加成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "IP地址格式错误或已在白名单",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "添加失败",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "将IP地址从白名单中移除，该IP将重新受到安全检查",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "安全管理"
                ],
                "summary": "从白名单移除IP",
                "parameters": [
                    {
                        "type": "string",
                        "description": "IP地址（示例: 192.168.1.1）",
                        "name": "ip",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "移除成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "IP地址格式错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "IP不在白名单中",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "移除失败",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "controllers.BlockedIP": {
            "type": "object",
            "properties": {
                "blocked_at": {
                    "type": "integer",
                    "example": 1640995200
                },
                "duration": {
                    "type": "integer",
                    "example": 3600
                },
                "expires_at": {
                    "type": "integer",
                    "example": 1640998800
                },
                "ip": {
                    "type": "string",
                    "example": "192.168.1.100"
                },
                "reason": {
                    "type": "string",
                    "example": "恶意攻击"
                }
            }
        },
        "controllers.ConnectionStats": {
            "type": "object",
            "properties": {
                "active_connections": {
                    "type": "integer",
                    "example": 10
                },
                "ip_connections": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer",
                        "format": "int64"
                    }
                },
                "top_ips": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/controllers.IPStats"
                    }
                },
                "total_connections": {
                    "type": "integer",
                    "example": 25
                }
            }
        },
        "controllers.IPStats": {
            "type": "object",
            "properties": {
                "ip": {
                    "type": "string",
                    "example": "192.168.1.1"
                },
                "is_blocked": {
                    "type": "boolean",
                    "example": false
                },
                "is_whitelisted": {
                    "type": "boolean",
                    "example": false
                },
                "last_access": {
                    "type": "integer",
                    "example": 1640995200
                },
                "request_count": {
                    "type": "integer",
                    "example": 150
                }
            }
        },
        "controllers.SecurityStatus": {
            "type": "object",
            "properties": {
                "config": {},
                "rate_limiter_active": {
                    "type": "boolean",
                    "example": true
                },
                "threat_level": {
                    "type": "string",
                    "example": "low"
                },
                "uptime": {
                    "type": "integer",
                    "example": 3600
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务状态码",
                    "allOf": [
                        {
                            "$ref": "#/definitions/response.ResponseCode"
                        }
                    ]
                },
                "data": {
                    "description": "数据"
                },
                "msg": {
                    "description": "消息",
                    "type": "string"
                },
                "timestamp": {
                    "description": "时间戳",
                    "type": "integer"
                }
            }
        },
        "response.ResponseCode": {
            "type": "integer",
            "enum": [
                200,
                400,
                401,
                403,
                404,
                405,
                409,
                422,
                429,
                500,
                501,
                502,
                503
            ],
            "x-enum-varnames": [
                "CodeSuccess",
                "CodeBadRequest",
                "CodeUnauthorized",
                "CodeForbidden",
                "CodeNotFound",
                "CodeMethodNotAllowed",
                "CodeConflict",
                "CodeUnprocessableEntity",
                "CodeTooManyRequests",
                "CodeInternalServerError",
                "CodeNotImplemented",
                "CodeBadGateway",
                "CodeServiceUnavailable"
            ]
        }
    },
...
}